requirements.txt
absl-py==1.1.0
albumentations==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2023.7.22
colorama==0.4.5
cycler==0.11.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.19.3
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.5
oauthlib==3.2.0
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-macos==2.9.0
tensorflow-metal==0.5.0
tensorflow-estimator==2.9.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
